LayoutTests/platform/blackberry/http/tests/loading/file-url-with-query-string-and-anchor.html

<html>
<head>
<script>
if (window.layoutTestController) {
    layoutTestController.dumpAsText();
}
</script>
</head>
<body>
<p>This test can be used to verify that loading a file URL with a query string and anchor works.</p>
<div id="testResult">Failure</div><br>
<iframe src="resources/file-url-with-query-string-and-anchor.html?q=dummy#anchor"></iframe>
</body>
</html>
